sem5/omis/lab2/requirements.txt
fastapi
uvicorn
jinja2
starlette
itsdangerous
python-multipart
pydantic
psycopg2-binary
python-dotenv
